demo/weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.71.0",
    "dashscope>=1.24.6",
    "httpx>=0.28.1",
    "mcp[cli]>=1.18.0",
    "python-dotenv>=1.1.1",
]
